fix(doctor): fix atp bug doctor app
